Servicio_de_comidas/requirements.txt
Flask==2.3.3
Flask-RESTx==1.1.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
werkzeug<3.0.0
